rewrite and rename to mkmatter
