+++ .rake_tasks~
@@ -1,0 +1,6 @@
+build
+clean
+clobber
+install
+install:local
+release[remote]
