+++ bin/mkmatter
@@ -1,7 +1,7 @@
 #!/usr/bin/ruby
 require 'highline'
-require 'jekyllposter/questions'
-require 'jekyllposter/answers'
+require 'mkmatter/questions'
+require 'mkmatter/answers'
 require 'slugity/extend_string'
 require 'find'
 Signal.trap('INT') do
@@ -25,11 +25,11 @@ def main
   end
   case layout
     when :post
-      @questions = Jekyllposter::Questions::Post.new(@hl).ask
+      @questions = Mkmatter::Questions::Post.new(@hl).ask
     when :page
-      @questions = Jekyllposter::Questions::Page.new(@hl).ask
+      @questions = Mkmatter::Questions::Page.new(@hl).ask
   end
-  @answers = Jekyllposter::Answers.new(draft, layout, @questions)
+  @answers = Mkmatter::Answers.new(draft, layout, @questions)
   
   @stdout = @hl.agree 'Would you like to output to stdout?', true
