--- lib/jekyllposter.rb
@@ -1,4 +1,0 @@
-require 'jekyllposter/version'
-require 'jekyllposter/common'
-require 'jekyllposter/questions'
-require 'jekyllposter/answers'
