+++ lib/mkmatter.rb
@@ -1,0 +1,4 @@
+require 'mkmatter/version'
+require 'mkmatter/common'
+require 'mkmatter/questions'
+require 'mkmatter/answers'
