+++ lib/mkmatter/common.rb
