+++ lib/mkmatter/version.rb
@@ -1,3 +1,3 @@
-module Jekyllposter
+module Mkmatter
   VERSION = '1.0.10'
 end
